refactor: move and route detax via `AntwebFormatter`
